redis cache

+++ backend/todo-main/TodoHub.Main/TodoHub.Main.API/Controllers/UsersController.cs
@@ -5,12 +5,12 @@ using TodoHub.Main.Core.Interfaces;
 
 namespace TodoHub.Main.API.Controllers
 {
-    // это контроллер собственно уже для управления users
+    // users controller
     [Route("/")]
     [ApiController]
     public class UsersController : ControllerBase
     {
-        // принимает сервис
+        
         private readonly IUserService _userService;
 
         public UsersController(IUserService userService)
@@ -18,52 +18,64 @@ namespace TodoHub.Main.API.Controllers
             _userService = userService;
         }
 
-        // тут чето типа загрузки текущего профиля человека
+
         [HttpGet("profile")]
         [Authorize]
         public async Task<IActionResult> GetProfile()
         {
-            // берем Id из токена(из клейма)
+            // take the ID from the token
             var userIdClaim = User.FindFirst("UserId")?.Value;
-            // нету ? значит не авторизован
+             
             if (userIdClaim is null)
                 return Unauthorized();
-            // пробуем запарсить
+            // parse
             if (!Guid.TryParse(userIdClaim, out var userId))
                 return BadRequest("Invalid user id in token");
-            // ищем это айди
+            // search for id
             var user = await _userService.GetMe(userId);
             if (!user.Success)
                 return NotFound();
-            // возвращаем если нашли
+            // return
             return Ok(user);
         }
 
-        // разрешенно только если пользователь админ
+        // only allowed if the user is an administrator
         [Authorize(Roles = "Admin")]
         [HttpGet("users")]
         public async Task<ActionResult<List<UserDTO>>> GetUsers()
         {
+            var userIdClaim = User.FindFirst("UserId")?.Value;
+
+            if (userIdClaim is null)
+                return Unauthorized();
             var users = await _userService.GetUsersAsync();
             return Ok(users);
         }
 
-        // разрешенно только если пользователь админ
+        // only allowed if the user is an administrator
         [Authorize(Roles = "Admin")]
         [HttpGet("profile/{id}")]
         public async Task<ActionResult<UserDTO>> GetUserById(Guid id)
         {
+            var userIdClaim = User.FindFirst("UserId")?.Value;
+
+            if (userIdClaim is null)
+                return Unauthorized();
             var user = await _userService.GetUserByIdAsync(id);
             return Ok(user);
         }
 
 
 
-        // разрешенно только если пользователь админ
+        // only allowed if the user is an administrator
         [Authorize(Roles = "Admin")]
         [HttpDelete("user/delete/{id}")]
         public async Task<IActionResult> DeleteUser(Guid id)
         {
+            var userIdClaim = User.FindFirst("UserId")?.Value;
+
+            if (userIdClaim is null)
+                return Unauthorized();
             var result = await _userService.DeleteUserAsync(id);
             if (!result.Success)
             {
